Update: Collections Division.
Changes were made to the database model so the results were separated into 4 different collections; now the peaks, tfbinding and authorsData are kept in their own collection.
The function to determine the closest gene of the peaks and tfbinding has been added (it is still in development but there is already a version).

Phill Betancourt ☕