Update: Closest Genes.
The method to obtain the closest genes has been added and improved.
It was optimized as much as possible by separating the part that calculates the intervals so that it is done only once and not in each iteration.
At the moment, it was not possible to optimize the queries to the database since a different query is made for each case, but improvement is planned in the future.

Phill Betancourt ☕