Update: Peaks and Sites correction.
The sites were being taken from another place that was not, in addition they were being assigned a property that should go in the collection of peaks.
Now the sites are processed first so that they are later registered in the peaks, these changes have improved program performance.
Adjustments were made to the collection of datasets (metadata) in the properties linkedDatasets and sample.

Phill Betancourt ☕